+++ attributes/project.rb
@@ -1,7 +1,7 @@
 #
 # These attributes, which may be set in more than one cookbook, are intended to be set
 # once at project level for a given role and control the behaviour of multiple recipes.
-# 
+#
 # Author:: Andrew Coulton(<andrew@ingenerator.com>)
 # Cookbook Name:: ingenerator-php
 # Attribute:: project